+++ job_source_agent/providers/__init__.py
@@ -1,0 +1,10 @@
+from .base import AdapterResult, JobBoard, JobCandidate, JobQuery, ProviderAdapter
+
+__all__ = [
+    "AdapterResult",
+    "JobBoard",
+    "JobCandidate",
+    "JobQuery",
+    "ProviderAdapter",
+]
+